Completed Assignment
grid-get-next-robot.ss now uses A* to calculate an optimal path from the
robot to the goal, and then uses the first node along that path to
verify the potential nodes in a static evaluator. While we are not
allowed to write to the grid directly, according to the rules, there's
nothing preventing us from copying the grid to a virtual copy and
writing to that.